+++ cucumber-expressions/CHANGELOG.md
@@ -20,7 +20,9 @@ N/A
 N/A
 
 ### Fixed
-N/A
+* Throw an error when `ParameterType` regexps have flags. Flags are not allowed because only the source
+  of the regexp is used to build a new regexp for the entire Cucumber Expression. See
+  [#308](https://github.com/cucumber/cucumber/issues/308). ([aslakhellesoy])
 
 ## [5.0.4]